+++ src/tool/subcommands/api_cmd/generate_test_snapshot.rs
@@ -204,10 +204,7 @@ where
     }
 }
 
-impl<T> ReadOpsTrackingStore<T>
-where
-    T: Blockstore + SettingsStore,
-{
+impl<T> ReadOpsTrackingStore<T> {
     pub fn new(inner: T) -> Self {
         Self {
             inner,