+++ CHANGELOG.md
@@ -33,11 +33,15 @@
 
 ### Changed
 
+- [#6821](https://github.com/ChainSafe/forest/pull/6821): Added message receipt size and event size to `forest-tool archive info` output.
+
 ### Removed
 
 ### Fixed
 
 - [#6707](https://github.com/ChainSafe/forest/issues/6707): Added missing `GoldenWeek` network upgrade entry in `Filecoin.StateGetNetworkParams` RPC method.
+
+- [#6817](https://github.com/ChainSafe/forest/pull/6817): Fixed `StateSearchMsg` to return null instead of error when not found to match Lotus behaviour.
 
 ## Forest v0.32.4 "Mild Inconvenience"